Major iteration on the post and the data.
I think the basic structure of the post is good now.
I need to regenerate the data with more runs to make it more
appealing.  I also think some of the names of the factors are
obscure, and should be fixed.

However, commiting the current state feels like a wise thing.